things
admin & landlord: added deletion procedure

updated clean_properties.csv

+++ main/website/routers/admin_views.py
@@ -10,23 +10,36 @@ from ..models import Property, PropertyImages, Notifications
 
 admin = Blueprint('admin', __name__)
 
-@admin.route("/")
+@admin.route("/unapproved_properties")
 @login_required
 @admin_required
 def admin_panel():
-    unapproved_properties = db.paginate(Property.query.order_by(Property.created_at.desc()).filter_by(is_approved=0),
+    properties = db.paginate(Property.query.order_by(Property.created_at.desc()).filter_by(is_approved=0),
                                         per_page=5)
     if request.args.get('page'):
-        unapproved_properties = db.paginate(
+        properties = db.paginate(
             Property.query.order_by(Property.created_at.desc()).filter_by(is_approved=0), per_page=5,
             page=int(request.args.get('page')))
 
-    return render_template("admin_panel_page.html", user=current_user, unapproved_properties=unapproved_properties)
+    return render_template("admin_panel_page.html", user=current_user, properties=properties, unapproved=1, pending_deletion=0)
 
-@admin.route("/manage_property_listing/property-<int:property_id>", methods=["GET", "POST"])
+@admin.route("/pending_deletion")
 @login_required
 @admin_required
-def manage_approval(property_id):
+def admin_panel_pending_deletion():
+    properties = db.paginate(Property.query.order_by(Property.created_at.desc()).filter_by(is_pending_deletion=1),
+                                        per_page=5)
+    if request.args.get('page'):
+        properties = db.paginate(
+            Property.query.order_by(Property.created_at.desc()).filter_by(is_pending_deletion=1), per_page=5,
+            page=int(request.args.get('page')))
+
+    return render_template("admin_panel_page.html", user=current_user, properties=properties, pending_deletion=1, unapproved=0)
+
+@admin.route("/manage_property_listing/property-<int:property_id>/<string:manage_type>", methods=["GET", "POST"])
+@login_required
+@admin_required
+def manage_approval(property_id, manage_type):
     form = forms.AdminPropertyViewForm()
     property = Property.query.filter_by(property_id=property_id).first()
 
@@ -94,69 +107,33 @@ def manage_approval(property_id):
         flash(f"Property {property_id} rejected", "success")
         return redirect(url_for("admin.admin_panel"))
 
-    return render_template("admin_manage_properties_page.html", user=current_user, property=property, form=form)
+    # delete property
+    elif form.validate_on_submit() and form.delete_field:
+        Notifications.new_notification(property.user_id, "Property Deleted", f"Your request for property '{property.property_name}' to be deleted has been approved.")
+        # delete the images from database
+        current_image_url_list = PropertyImages.reject_property_images(property_id)
 
+        # delete the property from database
+        approval_document = Property.reject_property(property_id)
 
-# replaced with manage_approval(property_id)
-@admin.route("/manage_approval_document", methods=["GET", "POST"])
-@login_required
-@admin_required
-def manage_approval_document():
-    form = forms.ManageApprovalForm()
+        # delete the images
+        for image in current_image_url_list:
+            old_image_file_path = os.path.join(current_app.config['IMAGE_UPLOAD_FOLDER'], image.strip())
+            if os.path.exists(old_image_file_path):
+                # delete the image from folder
+                os.remove(old_image_file_path)
 
-    unapproved_properties = Property.query.filter_by(is_approved=0).all()
-    list_l = []
-    for i in unapproved_properties:
-        list_l.append(i.property_id)
-    flash(f"Unapproved properties: {list_l}")
+        # delete the old approval form
+        approval_form_file_path = os.path.join(current_app.config['APPROVAL_DOCUMENT_UPLOAD_FOLDER'],
+                                                approval_document)
+        if os.path.exists(approval_form_file_path):
+            os.remove(approval_form_file_path)
 
-    if form.validate_on_submit():
-        prop_id = form.property_id.data
-        selection = form.selection.data
+        flash(f"Property {property_id} deleted", "success")
+        return redirect(url_for("admin.admin_panel"))
 
-        if prop_id not in list_l:
-            flash("Invalid property ID", "error")
-            return render_template("manage_approval.html", user=current_user, form=form)
+    return render_template("admin_manage_properties_page.html", user=current_user, property=property, form=form, manage_type=manage_type)
 
-        if selection == "View documents":
-            filename = f"{prop_id}.pdf"
-            path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), 'storage',
-                                'approval_documents')
-            return send_from_directory(
-                directory=path,
-                path=filename,
-                as_attachment=False)
-
-        elif selection == "Yes":
-            Property.approve_property(prop_id)
-            flash("Property approved")
-
-        else:
-            # selection is "No"
-            # delete the images from database
-            current_image_url_list = PropertyImages.reject_property_images(prop_id)
-
-            # delete the proeprty from database
-            Property.reject_property(prop_id)
-
-            # delete the images
-            for image in current_image_url_list:
-                old_image_file_path = os.path.join(current_app.config['IMAGE_UPLOAD_FOLDER'], image.strip())
-                if os.path.exists(old_image_file_path):
-                    # delete the image from folder
-                    os.remove(old_image_file_path)
-
-            reformatted_filename = f"{prop_id}"
-            approval_form_file_path = os.path.join(current_app.config['APPROVAL_DOCUMENT_UPLOAD_FOLDER'],
-                                                   reformatted_filename + ".pdf")
-
-            # delete the old approval form
-            if os.path.exists(approval_form_file_path):
-                os.remove(approval_form_file_path)
-
-            flash("Property rejected, deleted from database", "error")
-
-    return render_template("manage_approval.html", user=current_user, form=form)
 
 @admin.route("/storage/<path:filename>")
 @login_required
